fast_base_login/app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
beanie==1.24.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
lazy-model==0.2.0
motor==3.3.2
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pymongo==4.6.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
toml==0.10.2
typing_extensions==4.9.0
uvicorn==0.25.0
